requirements.txt
asgiref==3.3.1
Django==3.1.7
django-restframework==0.0.1
environ==1.0
pytz==2021.1
sqlparse==0.4.1
